+++ src/types.ts
@@ -1,4 +1,5 @@
 export type Config = {
   wrapKeyNamesSeparately: boolean;
   addSpacesAroundPlusSign: boolean;
+  replaceWithIcons: boolean;
 };