+++ src/validKeyNames.ts
@@ -1,12 +1,17 @@
 export const validKeyNames = [
   "cmd",
+  "⌘", // "\\u{2318}"
   "shift",
+  "⇧",
   "ctrl",
+  "^",
   "alt",
   "enter",
   "esc",
   "tab",
   "space",
+  "opt",
+  "⌥", // "\\u{2325}"
   "F1",
   "F2",
   "F3",
@@ -18,5 +23,7 @@ export const validKeyNames = [
   "F9",
   "F10",
   "F11",
-  "F12"
+  "F12",
+  "pageup",
+  "pagedown"
 ];